* commit '81d95efa1ad962e81653688c6ebc328594b5c43b':
  hyperspectral: change update_display plotting
  apd_confocal: disable h5 save
  single_particle_blink: disable spectrum_after
  trpl_scan: live display choices available in update_display
  thorelabs_powermeter_analog_readout: remove unnecessary import
  kinetic_spectra: save full CCD images and wavelength array
  acton_spec: New COM port
  PLE updates
  apd_optimizer: add save_data setting
  apd_counter: allow for user input of input_terminal
  power_wheel_arduino: modified hardware with new pins
  monochromator_sweep: connect graph to self.ui
  mcl_xyz_stage: use debug_mode LQ rather than debug flag
  crystaltech_aotf: update COM port for new computer\
  apd_confocal: stop apd optimizer before starting apd scan
  trpl scans: move plotting out of _run function
  andor_ccd_readout: fix a pyqtgraph problem with image display
  pypicoharp: T3 mode test
  pypicoharp: Move DLL next to code

# Conflicts:
#	hardware_components/acton_spec.py
#	hardware_components/apd_counter.py
#	measurement_components/apd_optimizer.py

+++ measurement_components/andor_ccd_readout.py
@@ -215,7 +215,7 @@ class AndorCCDReadout(Measurement):
     
     def update_display(self):
 
-        self.img_item.setImage(self.buffer_.T, autoLevels=False)
+        self.img_item.setImage(self.buffer_.astype(np.float32).T, autoLevels=False)
         self.hist_lut.imageChanged(autoLevel=True, autoRange=True)
 
         self.spec_plot_line.setData(self.wls, self.spectra_data)
